requirements.txt
python-decouple
telethon
cryptg
heroku3
psycopg2
gitpython
asyncio